requirements.txt
python-dotenv
pyrogram
tgcrypto
google-generativeai
ormar
asyncpg
psycopg2
aiosqlite